backend/requirements.txt
# Flask Web Framework
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Cors==4.0.1
gunicorn
# Data Processing
pandas==2.2.3
openpyxl==3.1.5
rapidfuzz==3.10.0  # Fast fuzzy string matching for intelligent column mapping
python-dateutil==2.9.0  # Flexible date parsing for smart date format detection
xlrd==2.0.1  # Excel date handling support

# Security & Authentication  
PyJWT==2.8.0
Werkzeug==3.1.3
python-dotenv==1.0.0

# Additional dependencies
blinker==1.9.0
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2 
# PostgreSQL Connector
psycopg2-binary